[jail_run] Use privileged for both real and effective gid. JB#51740
Firejail does not really try to differentiate between effective and
real group ids, and normalizes egid to be the same as gid before
executing application in sandbox.

So, to make privileged egid survive firejail launch, we must set both real
and effective gid to privileged. And to still make it possible for the
launched application itself to switch back to real gid, it is stored in
saved gid slot.

Signed-off-by: Simo Piiroinen <simo.piiroinen@jolla.com>