[jail_run] Use private-bin when launching an application. Fixes JB#52400
When private-bin is not used, also application binaries outside sandbox
are locked in and can't be replaced. Which then makes for example sw
upgrades impossible without first terminating all sandboxed applications.

Make sailjail apply private-bin option for the application it is about to
launch. It both increases the security and makes sandbox utilize copies
of binaries and thus unblocks sw upgrades.

Signed-off-by: Simo Piiroinen <simo.piiroinen@jolla.com>