[client] Implement sailjail client. JB#53868 OMP#JOLLA-45
Should be backwards compatible with sailjail binary, but does dbus ipc
with sailjail daemon for prompting and obtaining application
information.

Command line given to client is rejected if it does not match Exec
line from desktop file that also defines application permissions.

Executed application must be a readable and executable elf binary
file. Effectively this means scripts with implicitly defined
shell/interpreter are not allowed.

A launch notification is sent to lipstick prior to executing the
application, and if execution fails also launch cancellation
notification is sent.

Signed-off-by: Simo Piiroinen <simo.piiroinen@jolla.com>