[sailjail] Support org and app name with firejail templates. OMP#JOLLA-88 JB#54034
Save the organization and application name when parsing the permission
file. Pass these as OrganizationName and ApplicationName key templates
to firejail to be utilized when using, e.g., ${ApplicationName} macro
on a D-Bus name ownership rule to allow owning an application specific
D-Bus name.

To not to break compatibility with plugin API move org and app name
variables to private rules and add a getter for these. Use the already
defined SAILJAIL_KEY_{APPLICATION,ORGANIZATION}_NAME as key values as
well to be consistent. Constify the arg for jail_rules_cast() -
apparently G_CAST approves const pointers as well.

Organization name can have 'A-Za-z0-9._-' characters and cannot start
with a digit, nor it can have two consecutive dots or dot followed with
a digit. Allow to use hyphen in the org name as rDNS spec allows it.
Since the template names can be used anywhere this restriction to follow
only D-Bus name spec is too restrictive..

Application name can have 'A-Za-z0-9_' characters and cannot start with
a digit.

Both organization and application names must end with an alphanumeric
char.

If either, organization or application name fails the checks error
message is printed and sailjail quits.

Signed-off-by: Jussi Laakkonen <jussi.laakkonen@jolla.com>