[sailjaild] Remove stale data. JB#54095 OMP#JOLLA-107
Remove state data when user or application is removed. Either it is
detected via inotify or it has happened when daemon wasn't running. If
user is removed or has been removed, their respecive data file is
removed and usersettings instance destroyed. If application profile is
removed or has been removed, then the appsettings instance is destroyed
and usersettings are saved again to remove stale data.

Signed-off-by: Tomi Leppänen <tomi.leppanen@jolla.com>