[unit] Add use of the priv rules to fix test_rules.c. OMP#JOLLA-88 JB#54034
Use (and copy struct def) the JailRulesPriv also in tests to allow
G_CAST to operate properly when using jail_rules_restrict(). Application
and organization names are accessed via the JailRulesPriv and this
causes segfault in tests unless the JailRulesPriv is setup as well.

For this reason move the JailRulesPriv to jail_rules_p.h so in the
future changes dont break the tests.