[sailjaild] Add opt out option. JB#54498 OMP#JOLLA-178
Add opt out option for sandboxing. This allows app to opt out from
default sandboxing. The option should to be validated in the package if
the contained app should not be allowed to opt out. This adds "None"
mode for sandboxing. This mode is a hint for launcher to skip sailjail
but it's still possible to launch an app with sandboxing in none mode.
In that case it will start in compatibility mode with default profile.

Also add option to enable default profile so that "Compatibility" mode
is reported only when default profile is enabled.

Signed-off-by: Tomi Leppänen <tomi.leppanen@jolla.com>