[sailjail] Allow specifying binary without path. Fixes JB#55089
This relaxes checks for all apps and not just those run in compatibility
mode. If Exec line defines only application binary (+ arguments to app)
then it may be defined without path. In that case we require that the
invoked binary path points to a file in /usr/bin and basename matches to
Exec line's first word.

Rewrote the earlier implementation for legacy apps because it could have
mistakenly detected arguments for sailjail as the binary name. This
check is simpler and easier to read as it only requires to check the
first word of Exec line for binary name. Previously legacy apps could
have had sailjail wrapper although they weren't sandboxing aware.

Removed that check template's first word for field code test as it's
checked already against the binary to launch, which itself is also
checked for being an elf file.

Signed-off-by: Tomi Leppänen <tomi.leppanen@jolla.com>