[prompter] Close dialog if client disconnects. Fixes JB#54644
If client, like sailjailclient, disconnects from D-Bus after prompting
for permissions, we should drop that prompt and if it's the active one
close the dialog. Add watcher_t to support watching of D-Bus names.

Listen to unique names of the clients leaving the bus and go through all
of the invocations when that happens. To avoid race conditions this also
includes NameHasOwner check to ensure that the client hasn't left the
bus between making the call and creating watch.

The state machine gets a new state called WAIT. Instead of handling
waiting of results as part of PROMPT state, it has its own state.
Normally PROMPT will always transition to WAIT through callback that
sets prompt object path. After that WAIT will determine whether to
cancel the prompt or call wait on it.

Signed-off-by: Tomi Leppänen <tomi.leppanen@jolla.com>