[sailjail] Allow D-Bus activation when there are share methods JB#56277 OMP#JOLLA-529
3rd party applications may define share methods which in practice
requires D-Bus auto start. Use Desktop Entry/Exec value if there is no
X-Sailjail/ExecDBus value.

This will cause applications to open a window when they are not running
but they are introspected, however that is rare to happen for random
applications. Without this, allowing ExecDBus, the alternative issue is
that the application can not be brought to foreground at all when it's
running but doesn't have a window open.

Signed-off-by: Tomi Leppänen <tomi.leppanen@jolla.com>