[tests] Fix tests depending on BINDIR. Fixes JB#56961
Some tests expect that BINDIR points to /usr/bin which makes sense,
however in QA tests it was overridden to point to elsewhere and that
made some tests fail. Local unittests didn't redefine that. Avoid
redefining BINDIR as it's not necessary and also rename that BINDIR in
tests.xml.in to TESTBINDIR to avoid confusion.

Signed-off-by: Tomi Leppänen <tomi.leppanen@jolla.com>